fix type error in sign function

+++ npcl/ops/local.py
@@ -47,7 +47,7 @@ def divergence2d(px, py):
 
 
 def sign(x):
-    return ((x >= 0).astype(np.float32)-0.5)*2
+    return ((x >= 0).astype(np.float32)-np.float32(0.5))*2
 
 
 def soft_shrink(x, mu):
